pyproject.toml
[project]
name = "bank-multi-agent"
version = "0.1.0"
description = "High-Efficiency Multi-Agent Expert System for Bank Operations"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    # LangGraph and LangChain
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    # Prompty for prompt management
    "prompty>=0.1.0",
    # Observability (Mandatory)
    "langfuse>=2.0.0",
    # REST API
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    # Async support
    "aiosqlite>=0.20.0",
    "httpx>=0.27.0",
    # Utilities
    "python-dotenv>=1.0.0",
    "tenacity>=9.0.0",
    "pip>=25.3",
    "ruff>=0.14.11",
    "pytest>=9.0.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.13"
strict = true
